fix: change dob format on demoStudentData
