+++ src/components/democlass/student/AddStudent.jsx
@@ -106,7 +106,7 @@ const AddStudent = ({ isAddStudentModalOpen, setIsAddStudentModalOpen, avatars }
                           </span>
                         </div>
                         <input
-                          className="border-2 border-gray-400 w-full rounded-lg p-2 outline-inputOutlineClr" 
+                          className="border-2 border-gray-400 w-full bg-white rounded-lg p-2 outline-inputOutlineClr" 
                           type="date" 
                           id="dob" 
                           name="dob"
